Refactored profile and update profile into separate components
